Created Base Entity And Doctor
I have created base entity for my application and as well as a Doctor entity  for the  'doctors ' table. I have used @PrePersist to generate the  doctor code using the title and specialty of the doctor entity. @PostPersist is used for logging purposes.

And I have also used Auditing feature.