Updated enum operations
I have updated the enum operations that were causing the error.

I have solved 2 different issue in this commit. One of them was about Database <=> entity colum type mismatch.The  Database expected enum types for enum values while our application was trying to set enums as strings.

So I have created a new sql file and changed the column types as varchar. I have added constraint to use it like enum types.

The other issue was about Bean lifecycle and the usage of @ConditionalOnProperty . I have learned that when we use @ConditionalOnProperty for a bean, that bean is registered in the container at  bean definition time while @Value annotations are resolved at bean initialization time. this meant I couldn't receive the application.properties values from a bean that created with @ConditionalOnProperty annotation.

To fix this, I used the @Component annotation and injected the mediflow.validate.phone property to check whether the phone should be validated.