Created basics of doctor department table
The doctor_department table is a join table for the doctor ⇔ department relationship. I decided to model this relationship using a Join Entity so that it would be more controllable.

I have implemented the basics of this module. For example, adding and removing a doctor from a department, finding all departments with their doctor lists, and so on.

Now, I need to add dynamic queries, for which I will use specifications. I have created some simple specifications, but I am not sure whether they work correctly.

Additionally , I realized that I have to study  Join queries with many-to-many relations